# lib/api-zod/src/generated/types/productPricesInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductPricesInput {
  /** @pattern ^[1-9][0-9]*$ */
  priceM: string;
  /** @pattern ^[1-9][0-9]*$ */
  priceL: string;
}